chore: clean up warp bits
