+++ node/rest/Cargo.toml
@@ -58,9 +58,6 @@ version = "0.8"
 [dependencies.snarkvm]
 workspace = true
 
-[dependencies.serde_json]
-version = "1"
-
 [dependencies.time]
 version = "0.3"
