+++ node/rest/src/helpers/mod.rs
@@ -19,9 +19,3 @@ pub use auth::*;
 
 mod error;
 pub use error::*;
-
-mod middleware;
-pub use middleware::*;
-
-mod or_reject;
-pub use or_reject::*;
