--- node/rest/src/helpers/or_reject.rs
@@ -1,33 +1,0 @@
-// Copyright (C) 2019-2023 Aleo Systems Inc.
-// This file is part of the snarkOS library.
-
-// The snarkOS library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// The snarkOS library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
-
-use crate::RestError;
-
-use anyhow::Result;
-use warp::{reject, Rejection};
-
-/// A trait to unwrap a `Result` or `Reject`.
-pub trait OrReject<T> {
-    /// Returns the result if it is successful, otherwise returns a rejection.
-    fn or_reject(self) -> Result<T, Rejection>;
-}
-
-impl<T> OrReject<T> for anyhow::Result<T> {
-    /// Returns the result if it is successful, otherwise returns a rejection.
-    fn or_reject(self) -> Result<T, Rejection> {
-        self.map_err(|e| reject::custom(RestError::Request(e.to_string())))
-    }
-}
